// src/views/water-reserve/rain-water/rain-interface.tsx
export interface Props {
    id?: number | null | undefined;
    idTram?: number | null | undefined;
    nam?: string | null | undefined;
    thang1?: number | null | undefined;
    thang2?: number | null | undefined;
    thang3?: number | null | undefined;
    thang4?: number | null | undefined;
    thang5?: number | null | undefined;
    thang6?: number | null | undefined;
    thang7?: number | null | undefined;
    thang8?: number | null | undefined;
    thang9?: number | null | undefined;
    thang10?: number | null | undefined;
    thang11?: number | null | undefined;
    thang12?: number | null | undefined;
}

export interface StationState {
    id?: number | null | undefined;
    idHuyen?: number | null | undefined;
    idXa?: number | null | undefined;
    tenTram?: string |null;
    ngayBatDau?: number | null | undefined;
    ngayKetThuc?: number | null | undefined;
}
  